Add common dynamic value for "creatures that died under your control this turn"
