+++ Mage.Sets/src/mage/cards/c/CallousSellSword.java
@@ -3,8 +3,7 @@ package mage.cards.c;
 import mage.MageInt;
 import mage.abilities.Ability;
 import mage.abilities.common.EntersBattlefieldAbility;
-import mage.abilities.dynamicvalue.DynamicValue;
-import mage.abilities.effects.Effect;
+import mage.abilities.dynamicvalue.common.CreaturesYouControlDiedCount;
 import mage.abilities.effects.OneShotEffect;
 import mage.abilities.effects.common.DamageWithPowerFromOneToAnotherTargetEffect;
 import mage.abilities.effects.common.counter.AddCountersSourceEffect;
@@ -21,7 +20,6 @@ import mage.game.Game;
 import mage.game.permanent.Permanent;
 import mage.target.common.TargetAnyTarget;
 import mage.target.common.TargetControlledCreaturePermanent;
-import mage.watchers.common.CreaturesDiedWatcher;
 
 import java.util.UUID;
 
@@ -33,7 +31,7 @@ public final class CallousSellSword extends AdventureCard {
     private static final FilterAnyTarget filterSecondTarget = new FilterAnyTarget("any other target");
 
     private static final Hint hint = new ValueHint(
-            "Creatures that died under your control this turn", CallousSellSwordValue.instance
+            "Creatures that died under your control this turn", CreaturesYouControlDiedCount.instance
     );
 
     public CallousSellSword(UUID ownerId, CardSetInfo setInfo) {
@@ -48,7 +46,7 @@ public final class CallousSellSword extends AdventureCard {
         this.addAbility(new EntersBattlefieldAbility(
                 new AddCountersSourceEffect(
                         CounterType.P1P1.createInstance(0),
-                        CallousSellSwordValue.instance, true
+                        CreaturesYouControlDiedCount.instance, true
                 ).setText("with a +1/+1 counter on it for each creature that died under your control this turn.")
         ).addHint(hint));
 
@@ -69,32 +67,6 @@ public final class CallousSellSword extends AdventureCard {
     @Override
     public CallousSellSword copy() {
         return new CallousSellSword(this);
-    }
-}
-
-enum CallousSellSwordValue implements DynamicValue {
-    instance;
-
-    @Override
-    public int calculate(Game game, Ability sourceAbility, Effect effect) {
-        return game.getState()
-                .getWatcher(CreaturesDiedWatcher.class)
-                .getAmountOfCreaturesDiedThisTurnByController(sourceAbility.getControllerId());
-    }
-
-    @Override
-    public CallousSellSwordValue copy() {
-        return this;
-    }
-
-    @Override
-    public String getMessage() {
-        return "creature that died under your control this turn";
-    }
-
-    @Override
-    public String toString() {
-        return "1";
     }
 }
