+++ Mage.Sets/src/mage/cards/s/SeasonOfLoss.java
@@ -1,10 +1,9 @@
 package mage.cards.s;
 
-import mage.abilities.Ability;
 import mage.abilities.Mode;
 import mage.abilities.dynamicvalue.DynamicValue;
 import mage.abilities.dynamicvalue.common.CardsInControllerGraveyardCount;
-import mage.abilities.effects.Effect;
+import mage.abilities.dynamicvalue.common.CreaturesYouControlDiedCount;
 import mage.abilities.effects.common.DrawCardSourceControllerEffect;
 import mage.abilities.effects.common.LoseLifeOpponentsEffect;
 import mage.abilities.effects.common.SacrificeAllEffect;
@@ -13,8 +12,6 @@ import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
 import mage.filter.StaticFilters;
-import mage.game.Game;
-import mage.watchers.common.CreaturesDiedWatcher;
 
 import java.util.UUID;
 
@@ -37,7 +34,7 @@ public final class SeasonOfLoss extends CardImpl {
         this.getSpellAbility().getModes().getMode().withPawPrintValue(1);
 
         // {P}{P} -- Draw a card for each creature you controlled that died this turn.
-        Mode mode2 = new Mode(new DrawCardSourceControllerEffect(SeasonOfLossValue.instance));
+        Mode mode2 = new Mode(new DrawCardSourceControllerEffect(CreaturesYouControlDiedCount.instance));
         this.getSpellAbility().addMode(mode2.withPawPrintValue(2));
 
         // {P}{P}{P} -- Each opponent loses X life, where X is the number of creature cards in your graveyard.
@@ -55,32 +52,5 @@ public final class SeasonOfLoss extends CardImpl {
     @Override
     public SeasonOfLoss copy() {
         return new SeasonOfLoss(this);
-    }
-}
-
-// Based on CallousSellSwordValue
-enum SeasonOfLossValue implements DynamicValue {
-    instance;
-
-    @Override
-    public int calculate(Game game, Ability sourceAbility, Effect effect) {
-        return game.getState()
-                .getWatcher(CreaturesDiedWatcher.class)
-                .getAmountOfCreaturesDiedThisTurnByController(sourceAbility.getControllerId());
-    }
-
-    @Override
-    public SeasonOfLossValue copy() {
-        return this;
-    }
-
-    @Override
-    public String getMessage() {
-        return "creature that died under your control this turn";
-    }
-
-    @Override
-    public String toString() {
-        return "1";
     }
 }
